Refactor Global Policies component
