+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/GlobalComponents/PolicyConfiguringDrawer.tsx
