+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/GlobalComponents/PolicyList.tsx
@@ -28,7 +28,7 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core';
 import CONSTS from 'AppData/Constants';
 import type { Policy } from '../Types';
-import TabPanel from '../components/TabPanel';
+import TabPanel from '../Components/TabPanel';
 import CreatePolicy from '../CreatePolicy';
 
 const useStyles = makeStyles(() => ({
